chore: rename files with camelCase naming
