+++ src/collection/commands/list_indexes.ts
