+++ src/types/gridfs.ts
@@ -1,6 +1,6 @@
 import { Binary, ObjectId } from "../../deps.ts";
 import { Document, ReadPreference } from "../types.ts";
-import { ReadConcern, WriteConcern } from "../types/readWriteConcern.ts";
+import { ReadConcern, WriteConcern } from "../types/read_write_concern.ts";
 
 export type FileId = ObjectId;
