+++ src/utils/saslprep/mod.ts
@@ -1,8 +1,9 @@
+// deno-lint-ignore-file camelcase
 //ORIGINAL PROJECT AND LICENSE IN: https://github.com/chiefbiiko/saslprep
 //ORIGINAL PROJECT AND LICENSE IN: https://github.com/chiefbiiko/sparse-bitfield
 //ORIGINAL PROJECT AND LICENSE IN: https://github.com/chiefbiiko/memory-pager
 import { Bitfield } from "./deps.ts";
-import { loadCodePoints } from "./loadCodePoints.ts";
+import { loadCodePoints } from "./load_code_points.ts";
 
 const {
   unassigned_code_points,
